Break out time series query
